requirements.txt
flask
flask-sqlalchemy
gunicorn
werkzeug
psycopg2-binary
python-dotenv
Flask-Cors
